Add Error handling in test scripts

+++ test/unit/test.sh
@@ -11,6 +11,10 @@ reset=`tput sgr0`
 
 # Move to gtest bin
 working_dir1="$current_work_dir/../../build/test/unit/gtest/hcfft_Create_Destroy_Plan/bin/"
+if [ ! -d "$working_dir1" ]; then
+  echo "Please run script[test.sh] from hcfft/test/unit/"
+  exit
+fi
 cd $working_dir1
 
 #Gtest functions
@@ -23,7 +27,8 @@ Log_file="$working_dir1/gtestlog.txt"
 if grep -q FAILED "$Log_file";
 then
     echo "${red}hcfft_Create_destroy_Plan               ----- [ FAILED ]${reset}"
-else
+  elif grep -q PASSED "$Log_file";
+  then
     echo "${green}hcfft_Create_destroy_Plan             ----- [ PASSED ]${reset}"
     rm $working_dir1/gtestlog.txt
 fi
@@ -46,7 +51,8 @@ do
   if grep -q FAILED "$Log_file";
   then
       echo "${red}${testdirectories[$i]}               ----- [ FAILED ]${reset}"
-  else
+  elif grep -q PASSED "$Log_file";
+  then
       echo "${green}${testdirectories[$i]}             ----- [ PASSED ]${reset}"
       rm $working_dir1/gtestlog.txt
   fi
@@ -61,7 +67,8 @@ do
   if grep -q FAILED "$Log_file";
   then
       echo "${red}${testdirectories[$i]}_double               ----- [ FAILED ]${reset}"
-  else
+  elif grep -q PASSED "$Log_file";
+  then
       echo "${green}${testdirectories[$i]}_double           ----- [ PASSED ]${reset}"
       rm $working_dir1/gtestlog.txt
   fi
@@ -76,7 +83,8 @@ do
   if grep -q FAILED "$Log_file";
   then
       echo "${red}${testdirectories[$i]}_padding               ----- [ FAILED ]${reset}"
-  else
+  elif grep -q PASSED "$Log_file";
+  then
       echo "${green}${testdirectories[$i]}_padding           ----- [ PASSED ]${reset}"
       rm $working_dir1/gtestlog.txt
   fi
